requirements.txt
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
av==10.0.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.6.2
browser-cookie3==0.19.1
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
curl-cffi==0.5.6
cycler==0.11.0
decorator==5.1.1
einops==0.6.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fake-useragent==1.1.3
filelock==3.12.0
Flask==2.3.2
Flask-Cors==4.0.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
ftfy==6.1.1
fvcore==0.1.5.post20221221
gitdb==4.0.10
GitPython==3.1.31
grad-cam==1.4.8
h11==0.14.0
huggingface-hub==0.15.1
idna==3.4
imageio==2.28.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
iopath==0.1.10
ipython==8.12.2
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
lazy_loader==0.2
Levenshtein==0.21.1
lz4==4.3.2
MailGw-Temporary-Email==0.0.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mmcv==2.0.0
mmengine==0.7.4
multidict==6.0.4
names==0.3.0
networkx==3.1
nibabel==5.1.0
numpy==1.24.3
opencv-python==4.7.0.72
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.1
parameterized==0.9.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
portalocker==2.7.0
prompt-toolkit==3.0.38
protobuf==4.23.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.18.0
pydantic==2.0
pydantic_core==2.0.1
pydeck==0.8.1b0
pydicom==2.3.1
pydub==0.25.1
PyExecJS==1.5.1
Pygments==2.15.1
pymailtm==1.1.1
Pympler==1.0.1
pyparsing==3.0.9
PyPasser==0.0.5
pyperclip==1.8.2
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytorchvideo==0.1.5
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
random-password-generator==2.2.0
random-username==1.0.2
rapidfuzz==3.1.1
regex==2023.6.3
requests==2.30.0
retrying==1.3.4
rich==13.4.1
safetensors==0.3.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
selenium==4.10.0
six==1.16.0
sklearn==0.0.post5
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
SpeechRecognition==3.8.1
stack-data==0.6.2
streamlit==1.24.0
streamlit-chat @ https://github.com/AI-Yash/st-chat/archive/refs/pull/24/head.zip#sha256=018c7b60a8a200eb23394ea309ced518f36c446612f74c788c309526a124eeba
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.2
tls-client==0.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.1
torchvision==0.13.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
trio==0.22.1
trio-websocket==0.10.3
ttach==0.0.3
TwoCaptcha==0.0.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.2
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
websocket-client==1.6.1
Werkzeug==2.3.6
wsproto==1.2.0
xlrd==2.0.1
yacs==0.1.8
yapf==0.33.0
yarl==1.9.2
zipp==3.15.0
